Feat: [BND-1116] multi initiative copies (#143)
