+++ src/components/Dashboard/__tests__/DashboardDropdownMenu.test.tsx
@@ -17,28 +17,28 @@ describe('DashboardDropdownMenu', () => {
 
     it('renders the Menu button', () => {
         render(<DashboardDropdownMenu selectedSection="bonus" onSectionChange={onSectionChange} />);
-        expect(screen.getByText('dashboard.menu')).toBeInTheDocument();
+        expect(screen.getByText('common.dashboard.menu')).toBeInTheDocument();
     });
 
     it('opens the drawer when Menu button is clicked', () => {
         render(<DashboardDropdownMenu selectedSection="bonus" onSectionChange={onSectionChange} />);
-        fireEvent.click(screen.getByText('dashboard.menu'));
+        fireEvent.click(screen.getByText('common.dashboard.menu'));
         expect(screen.getByRole('presentation')).toBeInTheDocument();
     });
 
     it('calls onSectionChange when an item is clicked', () => {
         render(<DashboardDropdownMenu selectedSection="bonus" onSectionChange={onSectionChange} />);
-        fireEvent.click(screen.getByText('dashboard.menu'));
-        fireEvent.click(screen.getByText('dashboard.faq'));
+        fireEvent.click(screen.getByText('common.dashboard.menu'));
+        fireEvent.click(screen.getByText('common.dashboard.faq'));
         expect(onSectionChange).toHaveBeenCalledWith('faq');
     });
 
     it('closes the drawer when close icon is clicked', async () => {
         render(<DashboardDropdownMenu selectedSection="bonus" onSectionChange={onSectionChange} />);
-        fireEvent.click(screen.getByText('dashboard.menu'));
+        fireEvent.click(screen.getByText('common.dashboard.menu'));
         const closeBtn = screen.getAllByRole('button')[1];
         fireEvent.click(closeBtn);
 
-        await waitForElementToBeRemoved(() => screen.queryByText('dashboard.faq'));
+        await waitForElementToBeRemoved(() => screen.queryByText('common.dashboard.faq'));
     });
 });
