+++ src/components/Dashboard/__tests__/DetailBonusCard.test.tsx
@@ -30,10 +30,10 @@ describe('DetailBonusCard', () => {
       />
     );
 
-    expect(screen.getByText('dashboard.detailBonusSection.bonusDetail')).toBeInTheDocument();
+    expect(screen.getByText('common.dashboard.detailBonusSection.bonusDetail')).toBeInTheDocument();
     expect(screen.getByText('€100.00')).toBeInTheDocument();
     expect(screen.getByText('RSSLNZ85T10H501Z')).toBeInTheDocument();
-    expect(screen.getByText('dashboard.voucherStatus.ACTIVE')).toBeInTheDocument();
+    expect(screen.getByText('common.dashboard.voucherStatus.ACTIVE')).toBeInTheDocument();
   });
 
   test('renders default fiscal number when provided as dash', () => {
@@ -55,7 +55,7 @@ describe('DetailBonusCard', () => {
       />
     );
 
-    const statusChip = screen.getByText('dashboard.voucherStatus.ACTIVE');
+    const statusChip = screen.getByText('common.dashboard.voucherStatus.ACTIVE');
     expect(statusChip).toBeInTheDocument();
   });
 
@@ -72,7 +72,7 @@ describe('DetailBonusCard', () => {
       />
     );
 
-    expect(screen.getByText('dashboard.voucherStatus.EXPIRED')).toBeInTheDocument();
+    expect(screen.getByText('common.dashboard.voucherStatus.EXPIRED')).toBeInTheDocument();
   });
 
   test('renders correct status chip for USED voucher', () => {
@@ -88,7 +88,7 @@ describe('DetailBonusCard', () => {
       />
     );
 
-    expect(screen.getByText('dashboard.voucherStatus.USED')).toBeInTheDocument();
+    expect(screen.getByText('common.dashboard.voucherStatus.USED')).toBeInTheDocument();
   });
 
   test('renders correct status chip for EXPIRING voucher', () => {
@@ -104,7 +104,7 @@ describe('DetailBonusCard', () => {
       />
     );
 
-    expect(screen.getByText('dashboard.voucherStatus.EXPIRING')).toBeInTheDocument();
+    expect(screen.getByText('common.dashboard.voucherStatus.EXPIRING')).toBeInTheDocument();
   });
 
   test('formats different amount values correctly', () => {
@@ -131,10 +131,10 @@ describe('DetailBonusCard', () => {
       />
     );
 
-    expect(screen.getByText('dashboard.detailBonusSection.amount')).toBeInTheDocument();
-    expect(screen.getByText('dashboard.detailBonusSection.status')).toBeInTheDocument();
-    expect(screen.getByText('dashboard.detailBonusSection.voucherStartDate')).toBeInTheDocument();
-    expect(screen.getByText('dashboard.detailBonusSection.voucherEndDate')).toBeInTheDocument();
-    expect(screen.getByText('dashboard.detailBonusSection.fiscalNumber')).toBeInTheDocument();
+    expect(screen.getByText('common.dashboard.detailBonusSection.amount')).toBeInTheDocument();
+    expect(screen.getByText('common.dashboard.detailBonusSection.status')).toBeInTheDocument();
+    expect(screen.getByText('common.dashboard.detailBonusSection.voucherStartDate')).toBeInTheDocument();
+    expect(screen.getByText('common.dashboard.detailBonusSection.voucherEndDate')).toBeInTheDocument();
+    expect(screen.getByText('common.dashboard.detailBonusSection.fiscalNumber')).toBeInTheDocument();
   });
 });
