+++ src/components/LandingPage/__tests__/CustomLandingSection.test.tsx
@@ -3,15 +3,11 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import CustomLandingSection from '../CustomLandingSection';
 
-const mockGetInitiative = jest.fn();
-
 jest.mock('react-i18next', () => ({
   useTranslation: () => ({
     t: (key: string) => {
       const translations: Record<string, string> = {
         'landing.requestBonus': 'Request Bonus Desktop',
-        'landing.requestBonusMobile': 'Request Bonus Mobile',
-        'landing.requestBonusTest': 'Richiedi il Bonus Test',
         'landing.loginMethods': 'Use SPID or CIE',
         'landing.continueOnWeb': 'Continue on Web',
       };
@@ -22,10 +18,6 @@ jest.mock('react-i18next', () => ({
 
 jest.mock('../../../hooks/useIsMobile', () => ({
   useIsMobile: jest.fn(),
-}));
-
-jest.mock('../../../utils/env', () => ({
-  getInitiative: () => mockGetInitiative(),
 }));
 
 jest.mock('react-router-dom', () => {
@@ -50,7 +42,6 @@ describe('CustomLandingSection', () => {
 
   beforeEach(() => {
     AuthCtxMock.loginMock.mockClear();
-    mockGetInitiative.mockReturnValue('bonusvalore');
   });
 
   afterEach(() => {
@@ -65,23 +56,6 @@ describe('CustomLandingSection', () => {
     expect(screen.getByText('Request Bonus Desktop')).toBeInTheDocument();
     expect(screen.getByText('Use SPID or CIE')).toBeInTheDocument();
     expect(screen.getByRole('button', { name: 'Continue on Web' })).toBeInTheDocument();
-  });
-
-  it('renders mobile version of title when isMobile is true', () => {
-    useIsMobile.mockReturnValue(true);
-
-    render(<CustomLandingSection />);
-
-    expect(screen.getByText('Request Bonus Mobile')).toBeInTheDocument();
-  });
-
-  it('renders the Bonus Test title for the bonustest initiative', () => {
-    useIsMobile.mockReturnValue(false);
-    mockGetInitiative.mockReturnValue('bonustest');
-
-    render(<CustomLandingSection />);
-
-    expect(screen.getByText('Richiedi il Bonus Test')).toBeInTheDocument();
   });
 
   it('calls login on button click', async () => {
