+++ src/locale/it/bonusElettrodomestici2025/copy.ts
@@ -1,0 +1,229 @@
+const it = {
+  "commons": {
+    "email": "Indirizzo email",
+    "confirmEmail": "Conferma indirizzo email",
+    "back": "Indietro",
+    "continue": "Continua",
+    "invalidEmail": "Email non valida",
+    "requiredField": "Campo obbligatorio",
+    "emailMismatch": "Le email non coincidono",
+    "exit": "Esci",
+    "mandatoryField": "Campo obbligatorio",
+    "footer": {
+      "a11y": "Accessibilità",
+      "PagoPA": "PagoPA S.p.A",
+      "legalInfo": "Società per azioni con socio unico - capitale sociale di euro 1,000,000 interamente versato - sede legale in Roma, Piazza Colonna 370,\n CAP 00187 - N. di iscrizione a Registro Imprese di Roma, CF e P.IVA 15376371009",
+      "privacy": "Informativa Privacy",
+      "personalData": "Diritto alla protezione dei dati personali",
+      "termsAndConditions": "Termini e condizioni d'uso",
+      "legalInfoText": "<0>PagoPA S.p.A.</0> - Società per azioni con socio unico - Capitale sociale di euro 1,000,000 interamente versato - Sede legale in Roma, Piazza Colonna 370, <2/> CAP 00187 - N. di iscrizione a Registro Imprese di Roma, CF e P.IVA 15376371009",
+      "privacyPolicyLink": "Informativa Privacy",
+      "termsAndConditionLink": "Termini e condizioni d’uso del sito ",
+      "informationSecurityLink": "Sicurezza delle informazioni ",
+      "assistanceLink": "Assistenza ",
+      "preLoginLinks": {
+        "aboutUs": {
+          "links": {
+            "aboutUs": "PagoPA S.p.A.",
+            "media": "Media",
+            "workwithud": "Lavora con noi"
+          }
+        },
+        "resources": {
+          "title": "Risorse",
+          "links": {
+            "privacyPolicy": "Informativa Privacy",
+            "certifications": "Certificazioni",
+            "informationsecurity": "Sicurezza delle informazioni",
+            "protectionofpersonaldata": "Diritto alla protezione dei dati personali",
+            "cookies": "Preferenze Cookie",
+            "termsandconditions": "Termini e Condizioni",
+            "transparentcompany": "Società trasparente",
+            "disclosurePolicy": "Responsible Disclosure Policy",
+            "model231": "Modello 231"
+          }
+        },
+        "followUs": {
+          "title": "Seguici su"
+        },
+        "accessibility": "Accessibilità"
+      },
+      "postLoginLinks": {
+        "privacyPolicy": "Informativa Privacy",
+        "protectionofpersonaldata": "Diritto alla protezione dei dati personali",
+        "termsandconditions": "Termini e Condizioni",
+        "accessibility": "Accessibilità"
+      },
+    },
+    "header": {
+      "pagopaLinkLabel": "PagoPA S.p.A.",
+      "pagopaLinkAriaLabel": "Link: vai al sito di PagoPA S.p.A.",
+      "pagopaLinkTitle": "Sito di PagoPA S.p.A.",
+      "productTitle": "Bonus Elettrodomestici",
+    },
+    "retry": "Riprova",
+    "login": "Accedi",
+  },
+
+  "landing": {
+    "requestBonus": "Richiedi il Bonus Elettrodomestici online",
+    "requestBonusMobile": "Accedi al sito Bonus Elettrodomestici",
+    "loginMethods": "Accedi con SPID o CIE per richiedere il tuo bonus.",
+    "continueOnWeb": "Continua sul sito",
+    "withIO": "Con IO è tutto più semplice",
+    "descriptionWithIO": "Gestisci tutto in un unico posto: richiedi il bonus, ricevi l’esito della richiesta e usa il bonus. Più veloce, più comodo.",
+    "downloadIO": "Scarica IO",
+  },
+  "assistance": {
+    "title": "Hai bisogno di aiuto per Bonus Elettrodomestici?",
+    "subtitle": "Indica l’indirizzo email in cui desideri ricevere le risposte dell’assistenza. ",
+    "emailPlaceholder": "Inserisci l’indirizzo email",
+    "confirmEmailPlaceholder": "Conferma l’indirizzo email",
+    "prePolicy": "Proseguendo dichiari di avere letto la ",
+    "next": "Avanti"
+  },
+  "upcomingInitiative": {
+    "title": "Ci siamo quasi!",
+    "preSubtitle": "Potrai richiedere il ",
+    "boldSubtitle": "Bonus Elettrodomestici ",
+    "postSubtitle": "a partire dalle ",
+    "secondBold": "ore 7:00 del 18 novembre 2025. ",
+    "thirdSubtitle": "\nAncora un po’ di pazienza!",
+    "withIO": "Con IO è tutto più semplice",
+    "descriptionWithIO": "Gestisci tutto in un unico posto: richiedi il bonus, ricevi l’esito della richiesta e usa il bonus. Più veloce, più comodo.",
+    "downloadIO": "Scarica l'app"
+  },
+  "bonus": "Bonus Elettrodomestici",
+  "tos": {
+    "description": "È il contributo, erogato dal Ministero delle Imprese e del Made in Italy, per incentivare la <b>sostituzione di un elettrodomestico</b> con un modello ad alta efficienza energetica e promuovere la sostenibilità e la transizione energetica.",
+
+    "sideMenu": {
+      "element1": {
+        "title": "Chi può richiederlo?",
+        "description": "Tutti i <bold>maggiorenni residenti in Italia</bold> che acquistano un elettrodomestico nuovo e smaltiscono quello vecchio dello stesso tipo.<br/>Si può richiedere <bold>un solo bonus per ogni famiglia anagrafica</bold>."
+      },
+      "element2": {
+        "title": "Cosa offre?",
+        "description": "Uno sconto del 30% sul prezzo d’acquisto di un solo elettrodomestico, nuovo e ad alta efficienza, di una di queste categorie:<br/> lavatrici e lavasciuga, forni, cappe da cucina, lavastoviglie, asciugabiancheria, frigoriferi, congelatori e piani cottura.<br/> Lo sconto massimo che puoi ottenere è:",
+        "listItem1": "<bold>Fino a 100 euro</bold>",
+        "listItem2": "<bold>Fino a 200 euro</bold> se il tuo ISEE è inferiore a 25.000 €",
+        "subDescription": "Lo sconto non è cumulabile con altre agevolazioni, anche di tipo fiscale.",
+        "link": "Mostra prodotti compatibili",
+      },
+      "element3": {
+        "title": "Come richiederlo?",
+        "description": "Puoi richiederlo tramite questo sito web o dall’app IO, fino a che saranno disponibili i fondi.<br/>Una volta effettuate le verifiche, a seconda del canale usato, riceverai un’email con l’esito o un messaggio in app."
+      },
+      "element4": {
+        "title": "Come si usa?",
+        "description": "Puoi usare il bonus <bold>entro 15 giorni</bold> da quando ricevi il messaggio di conferma dell’attivazione.<br/><bold>Nei negozi fisici:</bold> Scarica il codice a barre in questo sito o genera il codice dall’app IO.<br/> <bold>Online:</bold> inserisci il codice generato in questo sito o usa l’app IO per generare e usare lo sconto.",
+        "link": "Mostra l’elenco dei negozi"
+      }
+    },
+
+    "postDescription": "<bold>Ricorda:</bold> quando acquisti, chiedi al venditore come smaltire il vecchio elettrodomestico. Lo smaltimento deve essere gestito gratuitamente dal venditore.",
+    "privacy_part1": "Puoi leggere",
+    "privacy_terms": "Termini e condizioni d’uso",
+    "privacy_part2": "e l'",
+    "privacy_policy": "Informativa Privacy",
+    "privacy_part3": "per conoscere le modalità di utilizzo e trattamento dei dati.",
+    "continue": "Continua",
+  },
+
+  "verifyRequirements": {
+    "exit": "Esci",
+    "title": "Verifica dei requisiti",
+    "description":
+      "Per accedere al Bonus Elettrodomestici, consulteremo altri enti per verificare se hai i requisiti previsti.",
+
+    "family": {
+      "title": "Famiglia anagrafica",
+      "description1":
+        "Per accedere all’iniziativa, comunichiamo direttamente con il Ministero dell’Interno per verificare questa informazione.",
+      "description2":
+        "In questo modo non devi produrre nessuna autocertificazione.",
+    },
+
+    "selfDeclaration": {
+      "title": "Autodichiarazione",
+      "description":
+        "Dichiaro che, ai sensi del D.P.R. 28 dicembre 2000, n. 445",
+      "switchLabel":
+        "Userò il bonus per l’acquisto di un elettrodomestico di classe energetica superiore destinato a sostituire un altro della stessa tipologia.",
+    },
+
+    "isee": {
+      "title": "Hai un ISEE 2025 in corso di validità?",
+      "description":
+        "Quando un ISEE è valido?",
+      "option.<25000": "Ho un ISEE inferiore a 25.000€",
+      "option.>=25000": "Ho un ISEE uguale o superiore a 25.000€",
+      "option.none": "Non ho un ISEE o preferisco non rispondere",
+      "hint.<25000": "Hai diritto fino a 200€. Verificheremo questa informazione con INPS",
+      "hint.>=25000": "Hai diritto fino a 100€",
+      "hint.none": "Hai diritto fino a 100€",
+    },
+
+    "error": "Seleziona un’opzione per continuare",
+    "selfDeclarationError": "Seleziona per continuare",
+    "submit": "Invia richiesta"
+  },
+
+  "insertEmail": {
+    "title": "Inserisci il tuo indirizzo email",
+    "description": "Ci servirà per comunicarti tutti gli aggiornamenti sul Bonus Elettrodomestici."
+  },
+  "FAQSection": {
+    "title": "Domande frequenti",
+    "firstAccordion": {
+      "title": "È obbligatorio smaltire un elettrodomestico?",
+      "description": "Sì, puoi usare il Bonus per acquistare un elettrodomestico nuovo solo se sostituisci un vecchio elettrodomestico della stessa tipologia e di classe energetica inferiore. Lo smaltimento è gratuito ed è a carico del venditore."
+    },
+    "secondAccordion": {
+      "title": "Se uso il Bonus Elettrodomestici posso usare anche altri sconti o benefici?",
+      "description": "No. Il bonus non è cumulabile con altre agevolazioni o altri benefici, anche di tipo fiscale."
+    },
+    "thirdAccordion": {
+      "title": "Quali elettrodomestici posso acquistare con il contributo?",
+      "description": "Puoi acquistare un elettrodomestico di una di queste categorie: lavatrici e lavasciuga, forni, cappe da cucina, lavastoviglie, asciugabiancheria, frigoriferi, congelatori e piani cottura, come previsto nel dettaglio dall’articolo 3, comma 1, lettera a), del Decreto interministeriale. Puoi consultare l’elenco degli elettrodomestici che fanno parte dell’iniziativa in questa lista."
+    },
+     "fourthAccordion": {
+      "title": "Ho usato il bonus per acquistare un elettrodomestico, ma vorrei effettuare un reso, cosa devo fare?",
+      "description": "Contatta il punto vendita presso il quale hai effettuato l’acquisto con il bonus per richiedere informazioni."
+    },
+    "fifthAccordion": {
+      "title": "Quando riceverò l’esito della richiesta?",
+      "description": "I tempi possono variare da qualche ora a qualche giorno, in base al volume delle richieste."
+    },
+    "sixthAccordion": {
+      "title": "Devo presentare qualche documento al momento dell'acquisto?",
+      "description": "Solo se hai richiesto il bonus tramite sito web, il venditore è tenuto a verificare il tuo codice fiscale."
+    },
+    "seventhAccordion": {
+      "title": "Cosa devo fare con l’elettrodomestico da smaltire?",
+      "description": "Concorda con il venditore come smaltire il vecchio elettrodomestico. Il ritiro e lo smaltimento sono gratuiti."
+    },
+    "eighthAccordion": {
+      "title": "Entro quanto tempo posso usare il Bonus Elettrodomestici?",
+      "description": "Entro 15 giorni solari dalla data di emissione. Il tuo bonus scade automaticamente alle 23:59 del quindicesimo giorno dall’emissione. Una volta scaduto, se saranno ancora disponibili i fondi, puoi comunque inviare una nuova richiesta, che seguirà l’ordine cronologico di arrivo."
+    },
+    "ninthAccordion": {
+      "title": "Come faccio a sapere se l’elettrodomestico da sostituire è di classe inferiore?",
+      "description": "Se hai acquistato dal 2021 in poi, puoi cercare il tuo elettrodomestico sul sito EPREL e verificare la classe energetica.\nSe il tuo modello non è presente su EPREL, significa che è stato prodotto prima dell’obbligo di registrazione (2021). In questo caso puoi chiedere informazioni al venditore."
+    },
+    "tenthAccordion": {
+      "title": "Come posso verificare lo stato della richiesta?",
+      "description": "Se hai l’app IO, puoi controllare lo stato dalla sezione “Servizi”. Se hai fatto richiesta dal sito bonuselettrodomestici.it, trovi lo stato del bonus nella sezione “Il tuo bonus”."
+    },
+    "eleventhAccordion": {
+      "title": "Mi è scaduto il Bonus, posso chiederlo di nuovo?",
+      "description": "Sì, puoi inviare una nuova richiesta, a condizione che nessun altro membro della tua famiglia anagrafica l'abbia già fatto.\nLa richiesta verrà presa in carico in base all'ordine di arrivo, compatibilmente con i fondi disponibili."
+    },
+    "twelfthAccordion": {
+      "title": "Posso usufruire del Bonus per un acquisto effettuato prima dell’uscita del bonus?",
+      "description": "No, il Bonus Elettrodomestici vale solo per acquisti futuri. Non è retroattivo."
+    }
+  }
+}
+
+export default it;
