+++ src/locale/it/common.ts
@@ -1,0 +1,105 @@
+const common = {
+  "exit": "Esci",
+  "click-day": {
+    "firstDescription": "Stiamo per collegarti.\nAttendi qualche secondo...",
+    "secondDescription": "Ancora un po’ di pazienza..."
+  },
+
+  "feedbackStates": {
+    "requestSubmitted": {
+      "title": "La tua richiesta è stata inviata!",
+      "description": "Controlla periodicamente la tua casella di posta.\nRiceverai tutti gli aggiornamenti all’indirizzo email che hai indicato.",
+      "subDescription": "Puoi verificare lo stato della tua richiesta da\nquesto sito o, se usi IO, dalla sezione Servizi dell’app."
+    },
+    "onEvaluation": {
+      "title": "La tua richiesta è in fase di valutazione",
+      "description": "Quando pronto, riceverai l’esito all’indirizzo email indicato."
+    },
+    "familyMember": {
+      "title": "Un'altra persona della tua\nfamiglia ha inviato la\nrichiesta",
+      "description": "È possibile inviare una sola richiesta per ogni\nfamiglia anagrafica."
+    },
+    "accessToken": {
+      "title": "Non è stato possibile accedere",
+      "description": "Si è verificato un problema durante l'accesso.\nRiprova tra qualche minuto.",
+    },
+    "ageRestriction": {
+      "title": "Non hai l’età minima richiesta",
+      "description": "Per richiedere il bonus, devi avere minimo 18 anni"
+    },
+    "waitingList": {
+      "title": "Sei in lista d’attesa",
+      "description": "La tua richiesta è in coda.\nTi contatteremo se saranno disponibili nuove\nrisorse."
+    },
+    "initiativeEnded": {
+      "title": "Le adesioni sono\nterminate",
+      "description": "Non è più possibile aderire all’iniziativa."
+    },
+    "budgetExhausted": {
+      "title": "I fondi messi a\ndisposizione dall’ente\nsono terminati",
+      "description": "Non è più possibile aderire all’iniziativa."
+    },
+    "sessionExpired": {
+      "title": "La tua sessione è scaduta",
+      "description": "Accedi per riprendere la sessione."
+    },
+    "unknownError": {
+      "title": "Qualcosa non ha funzionato",
+      "description": "Si è verificato un errore imprevisto.\nRiprova tra qualche minuto."
+    },
+    "technicalError": {
+      "title": "Non è possibile\ncompletare la richiesta\nper questa iniziativa",
+      "description": "Siamo già a lavoro per risolvere.\nRiprova tra un po’."
+    },
+    "tooManyRequests": {
+      "title": "Sembra che le cose\nvadano un po' a rilento",
+      "description": "Potreste essere in tanti in questo momento.\nRiprova tra qualche minuto."
+    }
+  },
+  "dashboard": {
+    "title": "Il tuo bonus",
+    "description": "Qui puoi visualizzare e scaricare il tuo bonus.",
+    "barcodeSection": {
+      "barcodeDescription": "codice a barre",
+      "downloadBarcode": "Scarica codice",
+      "showMerchants": "Mostra esercenti"
+    },
+    "detailBonusSection": {
+      "bonusDetail": "dettagli del Bonus",
+      "amount": "Importo disponibile",
+      "status": "Stato del bonus",
+      "voucherStartDate": "Valido fino al",
+      "voucherEndDate": "Bonus emesso il",
+      "fiscalNumber": "Codice Fiscale"
+    },
+    "voucherStatus": {
+      "ACTIVE": "Attivo",
+      "EXPIRING": "In scadenza",
+      "EXPIRED": "Scaduto",
+      "USED": "Utilizzato"
+    },
+    "operationsSection": {
+      "title": "Le tue operazioni",
+      "onboardingInitiative": "Adesione iniziativa"
+    },
+    "faq": "Domande frequenti",
+    "menu": "Menu"
+  },
+  "drawerDetail": {
+    "title": "Dettaglio dell’operazione",
+    "assetAmount": "Importo del bene",
+    "appliedDiscount": "Sconto riconosciuto",
+    "transactionInformation": "Informazioni sulla transazione",
+    "merchant": "Esercente",
+    "date": "Data",
+    "status": "Stato",
+    "idTransaction": "ID transazione",
+    "statusMap": {
+      "AUTHORIZED": "Autorizzata",
+      "CANCELLED": "Cancellata",
+      "REWARDED": "Assegnata"
+    }
+  }
+}
+
+export default common;
