+++ src/utils/initializeOneTrust.ts
@@ -1,0 +1,5 @@
+import { initializeCookieOneTrust } from './oneTrustLoader';
+
+initializeCookieOneTrust().catch((error: unknown) => {
+  console.log('Failed to initialize Cookie OneTrust: ', error);
+});
